Linux:To speed up the startup process by setting some modules to start in parallel.
The SPACEMIT_PARALLEL_BOOTING option can be enabled for use, but it is disabled by default.

This feature is designed to optimize the boot speed for the SPACEMIT platform.
Although it can improve boot time, it may potentially cause deadlocks during startup.
It is recommended to set the loglevel to 1 when using this feature.

Change-Id: I3bd76816253277182b44c5e999e9b9db3b3f4beb